Vlad (#38)
* signup done, database fixed

* comments added

* unnecessery files removed
